requirements.txt
discord.py
matplotlib
numpy
python-dotenv
seaborn
pandas
google-generativeai